Removing the logic to only train on high loss examples.
Is that logic something we really want? It does make it hard to see what our loss curves actually look like. It's also not clear why it's valuable, because examples with low loss won't get changed as much so it just probably complicates the training logic without offering clear gains.