First go at a 'hey koda' model.
For this model we do include koda background speech even though we do not
anticipate needing to say 'hey koda' over koda's speech. But the reasoning is
that it won't hurt and we really want a high recall for this model so it will
probably just help on that front.